requirements.txt
pathway 
requests 
pyhton-dotenv